+++ Gruntfile.js
@@ -25,14 +25,13 @@ module.exports = function (grunt) {
                     // Avoid `window` checking
                     browserifyOptions: {
                         standalone: 'dummyPlaceholder',
+                        basedir: __dirname, // Could try for consistency with any relative paths if still seeing https://github.com/axemclion/IndexedDBShim/issues/291 ; see also http://stackoverflow.com/a/33124979/271577
                         builtins: false, // No need to define in Node (if there are any)
-                        commondir: false, // No need in Node to look up to nearest common directory (if there are any)
                         browserField: false, // Avoid using `browser` entry in package.json
                         // https://github.com/substack/node-browserify/issues/1277#issuecomment-115198436
                         // One would think this wouldn't be necessary with `builts`: false
                         insertGlobalVars: {
                             process: function () { // Avoid having a non-Node polyfill added
-                                return;
                             }
                         }
                     }
@@ -64,14 +63,13 @@ module.exports = function (grunt) {
                     // Avoid `window` checking
                     browserifyOptions: {
                         standalone: 'dummyPlaceholder',
+                        basedir: __dirname, // Could try for consistency with any relative paths if still seeing https://github.com/axemclion/IndexedDBShim/issues/291 ; see also http://stackoverflow.com/a/33124979/271577
                         builtins: false, // No need to define in Node (if there are any)
-                        commondir: false, // No need in Node to look up to nearest common directory (if there are any)
                         browserField: false, // Avoid using `browser` entry in package.json
                         // https://github.com/substack/node-browserify/issues/1277#issuecomment-115198436
                         // One would think this wouldn't be necessary with `builts`: false
                         insertGlobalVars: {
                             process: function () { // Avoid having a non-Node polyfill added
-                                return;
                             }
                         }
                     }