+++ tests-mocha/IDBIndex/get-spec.js
@@ -1,4 +1,4 @@
-/* eslint-disable no-var */
+/* eslint-disable no-var, no-unused-expressions */
 describe('IDBIndex.get', function () {
     'use strict';