+++ tests-mocha/IDBObjectStore/add-put-spec.js
@@ -1,4 +1,4 @@
-/*eslint-disable no-var, spaced-comment*/
+/* eslint-disable no-var, spaced-comment, no-unused-expressions */
 /*********************************************************
  *     This file contains tests that are THE SAME for
  *       IDBObjectStore.add and IDBObjectStore.put