+++ tests-mocha/IDBTransaction/events-spec.js
@@ -51,7 +51,7 @@ describe('IDBTransaction events', function () {
 
             db.onerror = sinon.spy();
             tx.onerror = sinon.spy();
-            if (!env.browser.isFirefox) sinon.stub(window, 'onerror').returns(true);
+            if (!env.browser.isFirefox && !env.browser.isSafari) sinon.stub(window, 'onerror').returns(true);
 
             var store = tx.objectStore('inline');
             var errored = false;
@@ -65,7 +65,7 @@ describe('IDBTransaction events', function () {
                 expect(errored).equal(true);
                 sinon.assert.notCalled(db.onerror);             // <-- database.onerror NEVER fires
                 sinon.assert.notCalled(tx.onerror);             // <-- transaction.onerror NEVER fires
-                if (!env.browser.isFirefox) sinon.assert.notCalled(window.onerror);         // <-- window.onerror NEVER fires
+                if (!env.browser.isFirefox && !env.browser.isSafari) sinon.assert.notCalled(window.onerror);         // <-- window.onerror NEVER fires
             };
 
             setTimeout(function () {
@@ -73,11 +73,11 @@ describe('IDBTransaction events', function () {
                     expect(errored).equal(true);
                     sinon.assert.notCalled(db.onerror);         // <-- database.onerror NEVER fires
                     sinon.assert.notCalled(tx.onerror);         // <-- transaction.onerror NEVER fires
-                    if (!env.browser.isFirefox) sinon.assert.notCalled(window.onerror);     // <-- window.onerror NEVER fires
+                    if (!env.browser.isFirefox && !env.browser.isSafari) sinon.assert.notCalled(window.onerror);     // <-- window.onerror NEVER fires
                     db.close();
                     done();
                 } finally {
-                    if (!env.browser.isFirefox) window.onerror.restore();
+                    if (!env.browser.isFirefox && !env.browser.isSafari) window.onerror.restore();
                 }
             }, env.transactionDuration);
         });
@@ -104,7 +104,7 @@ describe('IDBTransaction events', function () {
                 sinon.assert.notCalled(tx.oncomplete);      // <-- transaction.oncomplete NEVER fires
                 sinon.assert.notCalled(tx.onerror);         // <-- transaction.onerror HAS NOT fired YET
                 sinon.assert.notCalled(db.onerror);         // <-- database.onerror HAS NOT fired YET
-                if (!env.browser.isFirefox) sinon.assert.notCalled(window.onerror);     // <-- window.onerror HAS NOT fired YET
+                if (!env.browser.isFirefox && !env.browser.isSafari) sinon.assert.notCalled(window.onerror);     // <-- window.onerror HAS NOT fired YET
             });
 
             // IDBTransaction events
@@ -113,8 +113,8 @@ describe('IDBTransaction events', function () {
                 sinon.assert.notCalled(tx.oncomplete);      // <-- transaction.oncomplete NEVER fires
                 sinon.assert.calledOnce(add.onerror);       // <-- request.onerror HAS fired
                 sinon.assert.notCalled(db.onerror);         // <-- database.onerror HAS NOT fired YET
-                if (!env.browser.isFirefox) sinon.assert.notCalled(window.onerror);     // <-- window.onerror HAS NOT fired YET
-                else e.preventDefault(); // Firefox reaches here unless preventDefault() is called (which will also cause our tests to fail)
+                if (!env.browser.isFirefox && !env.browser.isSafari) sinon.assert.notCalled(window.onerror);     // <-- window.onerror HAS NOT fired YET
+                else if (env.browser.isFirefox) e.preventDefault(); // Firefox reaches here unless preventDefault() is called (which will also cause our tests to fail)
             });
             tx.oncomplete = sinon.spy(function (e) {
                 db.close();
@@ -127,11 +127,11 @@ describe('IDBTransaction events', function () {
                 sinon.assert.notCalled(tx.oncomplete);      // <-- transaction.oncomplete NEVER fires
                 sinon.assert.calledOnce(add.onerror);       // <-- request.onerror HAS fired
                 sinon.assert.calledOnce(tx.onerror);        // <-- transaction.onerror HAS fired
-                if (!env.browser.isFirefox) sinon.assert.notCalled(window.onerror);     // <-- window.onerror HAS NOT fired YET
+                if (!env.browser.isFirefox && !env.browser.isSafari) sinon.assert.notCalled(window.onerror);     // <-- window.onerror HAS NOT fired YET
             });
 
             // Window events
-            if (!env.browser.isFirefox) sinon.stub(window, 'onerror').returns(true);
+            if (!env.browser.isFirefox && !env.browser.isSafari) sinon.stub(window, 'onerror').returns(true);
 
             setTimeout(function () {
                 try {
@@ -141,14 +141,14 @@ describe('IDBTransaction events', function () {
                     sinon.assert.calledOnce(tx.onerror);        // <-- transaction.onerror HAS fired
                     sinon.assert.calledOnce(db.onerror);        // <-- database.onerror HAS fired
 
-                    if (!env.browser.isFirefox) {
+                    if (!env.browser.isFirefox && !env.browser.isSafari) {
                         // Firefox fires window.onerror.  No other browser does.
                         sinon.assert.notCalled(window.onerror);
                     }
                     db.close();
                     done();
                 } finally {
-                    if (!env.browser.isFirefox) window.onerror.restore();
+                    if (!env.browser.isFirefox && !env.browser.isSafari) window.onerror.restore();
                 }
             }, env.transactionDuration);
         });
@@ -181,11 +181,11 @@ describe('IDBTransaction events', function () {
                 sinon.assert.notCalled(db.onerror);             // <-- database.onerror NEVER fires
                 sinon.assert.notCalled(tx.onerror);             // <-- transaction.onerror NEVER fires
                 sinon.assert.notCalled(tx.onabort);             // <-- transaction.abort NEVER fires
-                if (!env.browser.isFirefox) sinon.assert.notCalled(window.onerror);         // <-- window.onerror HAS NOT fired YET
-                if (!env.browser.isFirefox) throw new Error('Test Error');
+                if (!env.browser.isFirefox && !env.browser.isSafari) sinon.assert.notCalled(window.onerror);         // <-- window.onerror HAS NOT fired YET
+                if (!env.browser.isFirefox && !env.browser.isSafari) throw new Error('Test Error');
             });
 
-            if (!env.browser.isFirefox) {
+            if (!env.browser.isFirefox && !env.browser.isSafari) {
                 sinon.stub(window, 'onerror', function () {
                     sinon.assert.calledOnce(tx.oncomplete);         // <-- transaction.oncomplete HAS fired
                     sinon.assert.notCalled(db.onerror);             // <-- database.onerror NEVER fires
@@ -204,14 +204,14 @@ describe('IDBTransaction events', function () {
             setTimeout(function () {
                 try {
                     sinon.assert.calledOnce(tx.oncomplete);     // <-- transaction.oncomplete HAS fired
-                    if (!env.browser.isFirefox) sinon.assert.calledOnce(window.onerror);    // <-- window.onerror HAS fired
+                    if (!env.browser.isFirefox && !env.browser.isSafari) sinon.assert.calledOnce(window.onerror);    // <-- window.onerror HAS fired
                     sinon.assert.notCalled(db.onerror);         // <-- database.onerror NEVER fires
                     sinon.assert.notCalled(tx.onerror);         // <-- transaction.onerror NEVER fires
                     sinon.assert.notCalled(tx.onabort);         // <-- transaction.abort NEVER fires
                     db.close();
                     done();
                 } finally {
-                    if (!env.browser.isFirefox) window.onerror.restore();
+                    if (!env.browser.isFirefox && !env.browser.isSafari) window.onerror.restore();
                 }
             }, env.transactionDuration);
         });
@@ -243,11 +243,11 @@ describe('IDBTransaction events', function () {
                 sinon.assert.notCalled(db.onerror);             // <-- database.onerror NEVER fires
                 sinon.assert.notCalled(tx.onerror);             // <-- transaction.onerror NEVER fires
                 sinon.assert.notCalled(tx.oncomplete);          // <-- transaction.oncomplete NEVER fires
-                if (!env.browser.isFirefox) sinon.assert.notCalled(window.onerror);         // <-- window.onerror HAS NOT fired YET
-                if (!env.browser.isFirefox) throw new Error('Test Error');
+                if (!env.browser.isFirefox && !env.browser.isSafari) sinon.assert.notCalled(window.onerror);         // <-- window.onerror HAS NOT fired YET
+                if (!env.browser.isFirefox && !env.browser.isSafari) throw new Error('Test Error');
             });
 
-            if (!env.browser.isFirefox) {
+            if (!env.browser.isFirefox && !env.browser.isSafari) {
                 sinon.stub(window, 'onerror', function () {
                     sinon.assert.calledOnce(tx.onabort);            // <-- transaction.onabort HAS fired
                     sinon.assert.notCalled(db.onerror);             // <-- database.onerror NEVER fires
@@ -262,14 +262,14 @@ describe('IDBTransaction events', function () {
             setTimeout(function () {
                 try {
                     sinon.assert.calledOnce(tx.onabort);        // <-- transaction.onabort HAS fired
-                    if (!env.browser.isFirefox) sinon.assert.calledOnce(window.onerror);    // <-- window.onerror HAS fired
+                    if (!env.browser.isFirefox && !env.browser.isSafari) sinon.assert.calledOnce(window.onerror);    // <-- window.onerror HAS fired
                     sinon.assert.notCalled(db.onerror);         // <-- database.onerror NEVER fires
                     sinon.assert.notCalled(tx.onerror);         // <-- transaction.onerror NEVER fires
                     sinon.assert.notCalled(tx.oncomplete);      // <-- transaction.oncomplete NEVER fires
                     db.close();
                     done();
                 } finally {
-                    if (!env.browser.isFirefox) window.onerror.restore();
+                    if (!env.browser.isFirefox && !env.browser.isSafari) window.onerror.restore();
                 }
             }, env.transactionDuration);
         });